refactor(150512): :hammer: refatora dieta especial admin
